Add support for quotes.
If a single- or double-quote is after punctuation like period, question mark, etc, as well as comma, then include it in the match.

In fact, multiple quotes together will be matched after punctuation, thus catching a quote inside a quote.

While making this adjustment, also made it so that when spacing must be the cut-off point, it handles that better.